[Index Management] Branch data streams api tests lifecycle validation (#241105)
Fixes https://github.com/elastic/kibana/issues/187372

## Summary
The Data Stream serverless tests were failing in MKI for Security
because Security MKI enforces a default project-level retention. This PR
splits the ds test into three files: one for common tests, one for
serverless tests (`skipMKI`) and one for MKI (`['skipSvlWorkplaceAI',
'skipSvlOblt', 'skipSvlSearch', 'skipSvlSec']`). In MKI, the tests
assert is different if the project is Security or not.

Testes locally in MKI Security and Observability.

Flaky tes runner:
https://buildkite.com/elastic/kibana-flaky-test-suite-runner/builds/9657

(cherry picked from commit f8984ec6420af9b5f82efc9630708f4d10a73aeb)

# Conflicts:
#	x-pack/platform/test/serverless/api_integration/test_suites/index_management/datastreams/ds_common.ts
