+++ x-pack/platform/test/serverless/api_integration/test_suites/index_management/datastreams/index.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import type { FtrProviderContext } from '../../../ftr_provider_context';
+
+export default function ({ loadTestFile }: FtrProviderContext) {
+  describe('Data streams', function () {
+    loadTestFile(require.resolve('./ds_common'));
+    loadTestFile(require.resolve('./ds_mki'));
+    loadTestFile(require.resolve('./ds_serverless'));
+  });
+}
